Remove old PubEditorClient in favor of new one ✨
